+++ src/main/java/com/jakewharton/telecine/TelecineModule.java
@@ -1,0 +1,37 @@
+package com.jakewharton.telecine;
+
+import android.content.SharedPreferences;
+import dagger.Module;
+import dagger.Provides;
+import javax.inject.Singleton;
+
+import static android.content.Context.MODE_PRIVATE;
+
+@Module(injects = {
+    TelecineActivity.class, TelecineService.class
+})
+final class TelecineModule {
+  private static final String PREFERENCES_NAME = "telecine";
+  private static final boolean DEFAULT_SHOW_COUNTDOWN = true;
+  private static final boolean DEFAULT_HIDE_FROM_RECENTS = false;
+
+  private final TelecineApplication app;
+
+  TelecineModule(TelecineApplication app) {
+    this.app = app;
+  }
+
+  @Provides @Singleton SharedPreferences provideSharedPreferences() {
+    return app.getSharedPreferences(PREFERENCES_NAME, MODE_PRIVATE);
+  }
+
+  @Provides @Singleton @ShowCountdown BooleanPreference provideShowCountdownPreference(
+      SharedPreferences prefs) {
+    return new BooleanPreference(prefs, "show-countdown", DEFAULT_SHOW_COUNTDOWN);
+  }
+
+  @Provides @Singleton @HideFromRecents BooleanPreference provideHideFromRecentsPreference(
+      SharedPreferences prefs) {
+    return new BooleanPreference(prefs, "hide-from-recents", DEFAULT_HIDE_FROM_RECENTS);
+  }
+}