+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Change Log
 ==========
 
+0.2.0 *(In development)*
+------------------------
+
+ * Add preference toggle for disabling the three second countdown.
+ * Add preference toggle for removing the activity from the recents list.
+ * Remove separate processes for the activity and service.
+
+
 0.1.0 *(2014-10-25)*
 --------------------